refactor: useApplicableCoupon으로 쿠폰의 적용여부를 강조
