+++ src/hooks/coupon/index.ts
@@ -1,5 +1,5 @@
 export { default as useLoadCoupon } from './useLoadCoupon';
 export { default as useExpirationDate } from './useExpirationDate';
-export { default as useApplicable } from './useApplicable';
+export { default as useApplicableCoupon } from './useApplicableCoupon';
 export { default as useApplyCoupons } from './useApplyCoupons';
 export { default as useDiscount } from './useDiscount';
