+++ src/hooks/coupon/useApplicableCoupon.test.tsx
@@ -1,4 +1,4 @@
-import useApplicable from './useApplicable';
+import useApplicableCoupon from './useApplicableCoupon';
 import { renderHook } from '@testing-library/react';
 import { RecoilRoot } from 'recoil';
 import { selectedCartItems } from '@recoil/atoms';
@@ -10,7 +10,7 @@ import dayjs from '@utils/dayjs';
 describe('쿠폰 적용가능 여부 테스트', () => {
   describe('context: fixed coupon', () => {
     it('총액이 100000원 미만이면 쿠폰이 적용되지 않는다.', () => {
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot initializeState={({ set }) => set(selectedCartItems, mockCartItems)}>
             {children}
@@ -32,7 +32,7 @@ describe('쿠폰 적용가능 여부 테스트', () => {
           category: '고양이',
         },
       };
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot
             initializeState={({ set }) => set(selectedCartItems, [...mockCartItems, item])}
@@ -48,7 +48,7 @@ describe('쿠폰 적용가능 여부 테스트', () => {
 
   describe('context: buyXgetY coupon', () => {
     it('장바구니에 quantity가 2개 초과인 상품이 없다면 쿠폰이 적용되지 않는다.', () => {
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot initializeState={({ set }) => set(selectedCartItems, mockCartItems)}>
             {children}
@@ -70,7 +70,7 @@ describe('쿠폰 적용가능 여부 테스트', () => {
           category: '고양이',
         },
       };
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot
             initializeState={({ set }) => set(selectedCartItems, [...mockCartItems, item])}
@@ -86,7 +86,7 @@ describe('쿠폰 적용가능 여부 테스트', () => {
 
   describe('context: freeShipping coupon', () => {
     it('장바구니 총 금액이 freeShipping minimumAmount 기준보다 작을 때 쿠폰이 적용되지 않는다.', () => {
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot initializeState={({ set }) => set(selectedCartItems, [mockCartItems[0]])}>
             {children}
@@ -97,7 +97,7 @@ describe('쿠폰 적용가능 여부 테스트', () => {
       expect(result.current.isApplicable(freeShippingCoupon)).not.toBeTruthy();
     });
     it('장바구니 총 금액이 freeShipping minimumAmount 기준보다 크다면 쿠폰이 적용된다.', () => {
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot initializeState={({ set }) => set(selectedCartItems, mockCartItems)}>
             {children}
@@ -118,7 +118,7 @@ describe('쿠폰 적용가능 여부 테스트', () => {
       jest.useFakeTimers();
       jest.setSystemTime(dayjs('2024-05-23T03:00:00').toDate());
 
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot initializeState={({ set }) => set(selectedCartItems, mockCartItems)}>
             {children}
@@ -131,7 +131,7 @@ describe('쿠폰 적용가능 여부 테스트', () => {
     it('시작시간일 때 쿠폰이 적용된다.', () => {
       jest.useFakeTimers();
       jest.setSystemTime(dayjs('2024-05-23T04:00:00').toDate());
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot initializeState={({ set }) => set(selectedCartItems, mockCartItems)}>
             {children}
@@ -144,7 +144,7 @@ describe('쿠폰 적용가능 여부 테스트', () => {
     it('시작시간과 끝시간 사이일 때 쿠폰이 적용된다.', () => {
       jest.useFakeTimers();
       jest.setSystemTime(dayjs('2024-05-23T05:00:00').toDate());
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot initializeState={({ set }) => set(selectedCartItems, mockCartItems)}>
             {children}
@@ -157,7 +157,7 @@ describe('쿠폰 적용가능 여부 테스트', () => {
     it('끝시간일 때 쿠폰이 적용된다.', () => {
       jest.useFakeTimers();
       jest.setSystemTime(dayjs('2024-05-23T07:00:00').toDate());
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot initializeState={({ set }) => set(selectedCartItems, mockCartItems)}>
             {children}
@@ -170,7 +170,7 @@ describe('쿠폰 적용가능 여부 테스트', () => {
     it('끝시간 이후일 때 쿠폰이 적용되지 않는다.', () => {
       jest.useFakeTimers();
       jest.setSystemTime(dayjs('2024-05-23T08:00:00').toDate());
-      const { result } = renderHook(() => useApplicable(), {
+      const { result } = renderHook(() => useApplicableCoupon(), {
         wrapper: ({ children }) => (
           <RecoilRoot initializeState={({ set }) => set(selectedCartItems, mockCartItems)}>
             {children}
